tutoring-assistant/requirements.txt
langchain==0.0.160
openai==0.28.0
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0
tiktoken==0.8.0
pandas==2.2.3
docarray==0.40.0
wikipedia==1.4.0
# google-api-python-client==2.159.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.1
# langchain-community==0.0.38
# unstructured==0.16.14
duckduckgo-search==7.2.1